requirements.txt
Django==1.4.3
django-polymorphic==0.2
wsgiref==0.1.2
